extract AbsctractSelectStmt from views
