+++ celesta-sql/src/main/java/ru/curs/celesta/score/View.java
@@ -1,9 +1,8 @@
 package ru.curs.celesta.score;
 
 import java.io.*;
-import java.util.ArrayList;
+import java.util.Collections;
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -11,94 +10,72 @@ import java.util.Map;
  */
 public class View extends AbstractView {
 
-  Map<String, ViewColumnMeta<?>> columnTypes = null;
-  Expr whereCondition;
+    Map<String, ViewColumnMeta<?>> columnTypes = null;
 
-  View(GrainPart grainPart, String name) throws ParseException {
-    super(grainPart, name);
-    getGrain().addElement(this);
-  }
 
-  public View(GrainPart grainPart, String name, String sql) throws ParseException {
-    this(grainPart, name);
-    StringReader sr = new StringReader(sql);
-    CelestaParser parser = new CelestaParser(sr);
-    try {
-      try {
-        parser.select(this);
-      } finally {
-        sr.close();
-      }
-      finalizeParsing();
-    } catch (ParseException e) {
-      delete();
-      throw e;
+    View(GrainPart grainPart, String name) throws ParseException {
+        super(grainPart, name);
+        getGrain().addElement(this);
     }
-  }
 
-  @Override
-  String viewType() {
-    return "view";
-  }
-
-  @Override
-  void finalizeParsing() throws ParseException {
-    finalizeColumnsParsing();
-    finalizeWhereConditionParsing();
-    finalizeGroupByParsing();
-  }
-
-  void finalizeWhereConditionParsing() throws ParseException {
-    List<TableRef> t = new ArrayList<>(getTables().values());
-    if (whereCondition != null) {
-      whereCondition.resolveFieldRefs(t);
-      whereCondition.validateTypes();
+    public View(GrainPart grainPart, String name, String sql) throws ParseException {
+        this(grainPart, name);
+        StringReader sr = new StringReader(sql);
+        CelestaParser parser = new CelestaParser(sr);
+        try {
+            AbstractSelectStmt selectStmt = addSelectStatement();
+            try {
+                parser.select(selectStmt);
+            } finally {
+                sr.close();
+            }
+            selectStmt.finalizeParsing();
+            finalizeParsing();
+        } catch (ParseException e) {
+            delete();
+            throw e;
+        }
     }
-  }
 
-
-  @Override
-  void setWhereCondition(Expr whereCondition) throws ParseException {
-    if (whereCondition != null) {
-      List<TableRef> t = new ArrayList<>(getTables().values());
-      whereCondition.resolveFieldRefs(t);
-      whereCondition.assertType(ViewColumnType.LOGIC);
+    @Override
+    String viewType() {
+        return "view";
     }
-    this.whereCondition = whereCondition;
-  }
 
-  @Override
-  public Map<String, ViewColumnMeta<?>> getColumns() {
-    if (columnTypes == null) {
-      columnTypes = new LinkedHashMap<>();
-      for (Map.Entry<String, Expr> e : columns.entrySet()) {
-        ViewColumnMeta<?> meta = e.getValue().getMeta();
-        meta.setName(e.getKey());
-        columnTypes.put(e.getKey(), meta);
-      }
+    @Override
+    AbstractSelectStmt newSelectStatement() {
+        return new ViewSelectStmt(this);
     }
-    return columnTypes;
-  }
 
-  /**
-   * Creates CREATE VIEW script in different SQL dialects by using 'visitor' pattern.
-   *
-   * @param bw  stream that the saving is performed into
-   * @param gen  generator-visitor
-   * @throws IOException  error on writing to stream
-   */
-  public void createViewScript(PrintWriter bw, SQLGenerator gen) throws IOException {
-    bw.println(gen.preamble(this));
-    selectScript(bw, gen);
-  }
 
-  @Override
-  final void writeWherePart(PrintWriter bw, SQLGenerator gen) throws IOException {
-    if (whereCondition != null) {
-      bw.println();
-      bw.write("  where ");
-      bw.write(gen.generateSQL(whereCondition));
+    @Override
+    public Map<String, ViewColumnMeta<?>> getColumns() {
+        if (segments.size() > 0) {
+            if (columnTypes == null) {
+                columnTypes = new LinkedHashMap<>();
+                for (Map.Entry<String, Expr> e : segments.get(0).columns.entrySet()) {
+                    ViewColumnMeta<?> meta = e.getValue().getMeta();
+                    meta.setName(e.getKey());
+                    columnTypes.put(e.getKey(), meta);
+                }
+            }
+            return columnTypes;
+        } else {
+            return Collections.emptyMap();
+        }
     }
-  }
+
+    /**
+     * Creates CREATE VIEW script in different SQL dialects by using 'visitor' pattern.
+     *
+     * @param bw  stream that the saving is performed into
+     * @param gen generator-visitor
+     * @throws IOException error on writing to stream
+     */
+    public void createViewScript(PrintWriter bw, SQLGenerator gen) throws IOException {
+        bw.println(gen.preamble(this));
+        selectScript(bw, gen);
+    }
+
 
 }
