+++ pom.xml
@@ -11,7 +11,7 @@
     <url>https://corchestra.ru/wiki/index.php?title=Celesta</url>
 
     <modules>
-        <module>celesta-documentation</module>
+        <!--<module>celesta-documentation</module>-->
         <module>celesta-unit</module>
         <module>celesta-test</module>
         <module>celesta-sql</module>
